Keep unit tests in src/test during it split
Address review feedback on the integration-test split: the move was
too coarse for MilvusLoonWriterTest and missed a network test in
HelloSpec.

- Restore the three pure unit tests of MilvusLoonWriterTest (option
  parsing and the constructor validation-order regression) to src/test,
  leaving only the two MinIO tests in src/it as
  MilvusLoonWriterMinioIT (renamed so the two source sets do not share
  a class name)
- Drop HelloSpec's 'get segment info' test: it issued a real HTTP call
  to localhost:19530 with no assertion, so it stayed green with or
  without a server and violated the 'no external services' contract
  of 'sbt test'; keep the pure FloatConverter unit test

Unit tests still run via 'sbt test'; integration tests via
'sbt IntegrationTest/test'.
