+++ src/test/scala/example/HelloSpec.scala
@@ -1,8 +1,6 @@
 package example
 
 import com.zilliz.spark.connector.FloatConverter
-import com.zilliz.spark.connector.MilvusClient
-import com.zilliz.spark.connector.MilvusConnectionParams
 
 class HelloSpec extends munit.FunSuite {
   test("say hello") {
@@ -16,18 +14,5 @@ class HelloSpec extends munit.FunSuite {
     println(floatBytes)
     val float32FromBytes = FloatConverter.fromFloat16Bytes(floatBytes)
     println(float32FromBytes)
-  }
-
-  test("get segment info") {
-    val milvusClient = new MilvusClient(
-      MilvusConnectionParams(
-        uri = "http://localhost:19530",
-        token = "root:Milvus",
-        databaseName = "default"
-      )
-    )
-    val segmentInfo =
-      milvusClient.getSegmentInfo(461458951676887161L, 461458951677087167L)
-    println(segmentInfo)
   }
 }
